feat: WaitingQueueRedisRepository 구현체 작성 및 테스트

+++ src/main/java/com/example/hhplus/concert/domain/waitingqueue/model/WaitingQueueWithPosition.java
@@ -2,7 +2,7 @@ package com.example.hhplus.concert.domain.waitingqueue.model;
 
 public record WaitingQueueWithPosition(
     String uuid,
-    Integer position
+    Long position
 ) {
 
 }
